feat: polish weather animations and add 3D map experience

+++ supabase/migrations/001_initial_schema.sql
@@ -10,32 +10,44 @@
 
 -- Enable PostGIS
 CREATE EXTENSION IF NOT EXISTS postgis;
+
 CREATE EXTENSION IF NOT EXISTS postgis_topology;
 
 -- ────────────────────────────────────────────────────────────
 -- 1. STATIONS
 -- ────────────────────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS stations (
-  id           TEXT PRIMARY KEY,                      -- e.g. 'st-001'
-  slug         TEXT UNIQUE NOT NULL,                  -- URL-safe identifier
-  name         TEXT NOT NULL,                         -- display name
-  name_az      TEXT,                                  -- Azerbaijani name
-  line         TEXT NOT NULL CHECK (line IN ('red', 'green', 'purple')),
-  lat          DOUBLE PRECISION NOT NULL,
-  lon          DOUBLE PRECISION NOT NULL,
-  geom         GEOMETRY(Point, 4326),                 -- PostGIS point
-  station_type TEXT CHECK (station_type IN (
-                 'commuter', 'transfer', 'residential',
-                 'mixed', 'central', 'tourist', 'business'
-               )),
-  opened_year  INTEGER,
-  created_at   TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-  updated_at   TIMESTAMPTZ NOT NULL DEFAULT NOW()
+    id TEXT PRIMARY KEY, -- e.g. 'st-001'
+    slug TEXT UNIQUE NOT NULL, -- URL-safe identifier
+    name TEXT NOT NULL, -- display name
+    name_az TEXT, -- Azerbaijani name
+    line TEXT NOT NULL CHECK (
+        line IN ('red', 'green', 'purple')
+    ),
+    lat DOUBLE PRECISION NOT NULL,
+    lon DOUBLE PRECISION NOT NULL,
+    geom GEOMETRY (Point, 4326), -- PostGIS point
+    station_type TEXT CHECK (
+        station_type IN (
+            'commuter',
+            'transfer',
+            'residential',
+            'mixed',
+            'central',
+            'tourist',
+            'business'
+        )
+    ),
+    opened_year INTEGER,
+    created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+    updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 -- Geospatial index
 CREATE INDEX IF NOT EXISTS stations_geom_idx ON stations USING GIST (geom);
+
 CREATE INDEX IF NOT EXISTS stations_slug_idx ON stations (slug);
+
 CREATE INDEX IF NOT EXISTS stations_line_idx ON stations (line);
 
 -- Auto-populate geom from lat/lon
@@ -57,37 +69,39 @@ FOR EACH ROW EXECUTE FUNCTION stations_set_geom();
 --    This is the core dataset. NOT origin-destination data.
 -- ────────────────────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS station_daily_demand (
-  date         DATE NOT NULL,
-  station_id   TEXT NOT NULL REFERENCES stations(id) ON DELETE CASCADE,
-  entries      INTEGER NOT NULL CHECK (entries >= 0),
-  source_year  INTEGER NOT NULL,
-  created_at   TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-  PRIMARY KEY (date, station_id)
+    date DATE NOT NULL,
+    station_id TEXT NOT NULL REFERENCES stations (id) ON DELETE CASCADE,
+    entries INTEGER NOT NULL CHECK (entries >= 0),
+    source_year INTEGER NOT NULL,
+    created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+    PRIMARY KEY (date, station_id)
 );
 
 CREATE INDEX IF NOT EXISTS sdd_station_id_idx ON station_daily_demand (station_id);
+
 CREATE INDEX IF NOT EXISTS sdd_date_idx ON station_daily_demand (date DESC);
 
 -- ────────────────────────────────────────────────────────────
 -- 3. STATION EXITS
 -- ────────────────────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS station_exits (
-  id            TEXT PRIMARY KEY,
-  station_id    TEXT NOT NULL REFERENCES stations(id) ON DELETE CASCADE,
-  exit_no       INTEGER NOT NULL,
-  exit_label    TEXT NOT NULL,
-  address_text  TEXT,
-  lat           DOUBLE PRECISION NOT NULL,
-  lon           DOUBLE PRECISION NOT NULL,
-  geom          GEOMETRY(Point, 4326),
-  is_accessible BOOLEAN DEFAULT FALSE,
-  notes         TEXT,
-  created_at    TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-  updated_at    TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-  UNIQUE (station_id, exit_no)
+    id TEXT PRIMARY KEY,
+    station_id TEXT NOT NULL REFERENCES stations (id) ON DELETE CASCADE,
+    exit_no INTEGER NOT NULL,
+    exit_label TEXT NOT NULL,
+    address_text TEXT,
+    lat DOUBLE PRECISION NOT NULL,
+    lon DOUBLE PRECISION NOT NULL,
+    geom GEOMETRY (Point, 4326),
+    is_accessible BOOLEAN DEFAULT FALSE,
+    notes TEXT,
+    created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+    updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+    UNIQUE (station_id, exit_no)
 );
 
 CREATE INDEX IF NOT EXISTS exits_station_id_idx ON station_exits (station_id);
+
 CREATE INDEX IF NOT EXISTS exits_geom_idx ON station_exits USING GIST (geom);
 
 -- Auto-populate geom
@@ -107,15 +121,15 @@ FOR EACH ROW EXECUTE FUNCTION exits_set_geom();
 -- 4. WEATHER (CITY CENTER, 3-HOURLY)
 -- ────────────────────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS weather_city_center (
-  observed_at   TIMESTAMPTZ NOT NULL PRIMARY KEY,
-  temp_c        REAL,
-  feels_like_c  REAL,
-  humidity      REAL,         -- percent
-  wind_speed    REAL,         -- km/h
-  precipitation REAL,         -- mm
-  pressure      REAL,         -- hPa
-  condition     TEXT,
-  created_at    TIMESTAMPTZ NOT NULL DEFAULT NOW()
+    observed_at TIMESTAMPTZ NOT NULL PRIMARY KEY,
+    temp_c REAL,
+    feels_like_c REAL,
+    humidity REAL, -- percent
+    wind_speed REAL, -- km/h
+    precipitation REAL, -- mm
+    pressure REAL, -- hPa
+    condition TEXT,
+    created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 CREATE INDEX IF NOT EXISTS weather_observed_at_idx ON weather_city_center (observed_at DESC);
@@ -142,6 +156,7 @@ CREATE TABLE IF NOT EXISTS station_forecasts (
 );
 
 CREATE INDEX IF NOT EXISTS forecasts_station_id_idx ON station_forecasts (station_id);
+
 CREATE INDEX IF NOT EXISTS forecasts_date_idx ON station_forecasts (forecast_date DESC);
 
 -- ────────────────────────────────────────────────────────────
@@ -171,14 +186,14 @@ CREATE INDEX IF NOT EXISTS profiles_station_id_idx ON station_intraday_profiles 
 -- 7. STATION BASELINES
 -- ────────────────────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS station_baselines (
-  station_id     TEXT NOT NULL REFERENCES stations(id) ON DELETE CASCADE,
-  dow            INTEGER NOT NULL CHECK (dow BETWEEN 0 AND 6),  -- 0=Sunday
-  avg_entries    REAL,
-  median_entries REAL,
-  rolling_avg_7  REAL,
-  rolling_avg_30 REAL,
-  updated_at     TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-  PRIMARY KEY (station_id, dow)
+    station_id TEXT NOT NULL REFERENCES stations (id) ON DELETE CASCADE,
+    dow INTEGER NOT NULL CHECK (dow BETWEEN 0 AND 6), -- 0=Sunday
+    avg_entries REAL,
+    median_entries REAL,
+    rolling_avg_7 REAL,
+    rolling_avg_30 REAL,
+    updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+    PRIMARY KEY (station_id, dow)
 );
 
 CREATE INDEX IF NOT EXISTS baselines_station_idx ON station_baselines (station_id);
@@ -206,7 +221,7 @@ SELECT
   f.rider_facing_note,
   CASE
     WHEN f.baseline_entries > 0 THEN
-      ROUND(((f.predicted_entries - f.baseline_entries)::REAL / f.baseline_entries) * 100, 1)
+      ROUND((((f.predicted_entries - f.baseline_entries)::NUMERIC / f.baseline_entries::NUMERIC) * 100), 1)
     ELSE 0
   END AS delta_pct
 FROM stations s
